fixing ending and starting day
