+++ Assets/Grigor/Scripts/Gameplay/Interacting/World/MindPalaceBedInteractable.cs
@@ -20,6 +20,7 @@ namespace Grigor.Gameplay.World.Components
         private void MovePlayerToRoom()
         {
             roomManager.MovePlayerToRoom(RoomName.Start, characterRegistry.Player.transform.position);
+
             timeManager.OnDayStarted();
         }
     }
